+++ JourTrip/JourTrip/ViewController/ViewController.swift
@@ -12,24 +12,31 @@ import AlamofireObjectMapper
 
 class ViewController: BaseViewController {
     
-    var token = ""
+    let email = "alexa@example.com"
+    let password = "1234"
     
     override func viewDidLoad() {
         super.viewDidLoad()
         
-        // signup example
-        self.serviceProxy.register(email: "alex14@example.com", password: "1234") { (response, error) in
-            let userTest = response?.result
-            print(userTest ?? "")
+        // register example
+        self.serviceProxy.register(email: email, password: password) { (response, error) in
+            guard let email = response?.result?.email else {
+                    fatalError("email or password error")
+            }
+            self.login(email: email, password: self.password)
         }
-        
+    }
+    
+    fileprivate func login(email: String, password: String) {
         // login example
-        self.serviceProxy.login(email: "alex13@example.com", password: "1234") { (response, error) in
+        self.serviceProxy.login(email: email, password: password) { (response, error) in
             if let token = response?.token {
-                self.token = token
+                self.getSomeLocations(with: token)
             }
         }
-        
+    }
+    
+    fileprivate func getSomeLocations(with token: String) {
         // Few locations example
         var ayuntamientos: [Location] = []
